requirements.txt
flask
requests
python-dotenv
websocket-client
gunicorn
